Comments fixed

+++ tasks/views.py
@@ -1,6 +1,7 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response 
 from rest_framework import status
+from rest_framework.permissions import IsAuthenticated
 from django.shortcuts import get_object_or_404
 from rest_condition import Or
 
@@ -11,7 +12,7 @@ from tasks.serializers import TaskSerializer
 
 class TaskView(APIView):
 
-    permission_classes = [Or(IsAdmin, IsManager, IsEmployee)]
+    permission_classes = [IsAuthenticated, Or(IsAdmin, IsManager, IsEmployee)]
     
     def post(self,request):
         serializer = TaskSerializer(data=request.data)
@@ -26,13 +27,19 @@ class TaskView(APIView):
         if  not task_id:
             if role == CustomUser.ADMIN:
                 tasks=Tasks.objects.all()
-                serializer = TaskSerializer(tasks, many = True)
-                return Response(serializer.data)
-            
-            return Response({"message":"You are not authorized"}, status=status.HTTP_401_UNAUTHORIZED)
+            elif role == CustomUser.MANAGER:
+                tasks=Tasks.objects.filter(assigned_to=request.user).all()
+            elif role == CustomUser.EMPLOYEE:
+                tasks=Tasks.objects.filter(assigned_to=request.user).all()
+
+            serializer = TaskSerializer(tasks, many = True)
+            return Response(serializer.data)
         
         task=get_object_or_404(Tasks, id=task_id)
-        if role == CustomUser.ADMIN:
+        if not task.assigned_to:
+            return Response({"message":"The task is unassigned"})
+        
+        if role == CustomUser.ADMIN or (role == CustomUser.MANAGER and task.assigned_to.role == CustomUser.EMPLOYEE):
             serializer = TaskSerializer(task)
             return Response(serializer.data)
         else:
@@ -44,12 +51,12 @@ class TaskView(APIView):
     
     def patch(self, request, task_id):
         task = get_object_or_404(Tasks, id=task_id)
+        role = request.user.role
         
         if "assigned_to" in request.data:
-            if request.user.role == CustomUser.ADMIN:
-                user_id = request.data.get('assigned_to')
-                user = get_object_or_404(CustomUser, id=user_id)
-            
+            user_id = request.data.get('assigned_to')
+            user = get_object_or_404(CustomUser, id=user_id)
+            if role == CustomUser.ADMIN or (role == CustomUser.MANAGER and user.role == CustomUser.EMPLOYEE):
                 task.assigned_to = user
                 task.status = Tasks.TASK_STATUS[1][0]
                 task.save()
@@ -57,7 +64,7 @@ class TaskView(APIView):
             return Response({"message":"You are not authorized to assign task"}, status=status.HTTP_401_UNAUTHORIZED)
         
         task.status = request.data.get('status')
-        if task.assigned_to == request.user or request.user.role == CustomUser.ADMIN:
+        if task.assigned_to == request.user or role == CustomUser.ADMIN or ( role == CustomUser.MANAGER and task.assigned_to.role == CustomUser.EMPLOYEE):
             task.save()
             return Response({"message":"Task status updated successfully"},status=status.HTTP_200_OK)
         
@@ -65,5 +72,8 @@ class TaskView(APIView):
         
     def delete(self, request, task_id):
         task = get_object_or_404(Tasks, id=task_id)
+        role = request.user.role
+        if role == CustomUser.EMPLOYEE:
+            return Response({"message":"You are not Authorized"}, status=status.HTTP_401_UNAUTHORIZED)
         task.delete()
         return Response({"message":"Task deleted successfully"}, status=status.HTTP_200_OK)
